refactor: accounts와 members 패키지 분리
