+++ src/main/java/com/ssafy/solvedpick/accounts/dto/MembernameDTO.java
@@ -2,14 +2,14 @@ package com.ssafy.solvedpick.accounts.dto;
 
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
-import lombok.Data;
+import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
 @Getter
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
+@Builder
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+@AllArgsConstructor(access = AccessLevel.PROTECTED)
 public class MembernameDTO {
     private String username;
 }
