+++ src/main/java/com/ssafy/solvedpick/accounts/dto/SignInFormDTO.java
@@ -2,12 +2,12 @@ package com.ssafy.solvedpick.accounts.dto;
 
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
-import lombok.Data;
+import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
 @Getter
-@Data
+@Builder
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @AllArgsConstructor(access = AccessLevel.PROTECTED)
 public class SignInFormDTO {
