+++ src/main/java/com/ssafy/solvedpick/accounts/dto/SignupFormDTO.java
@@ -2,13 +2,13 @@ package com.ssafy.solvedpick.accounts.dto;
 
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
-import lombok.Data;
+import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
 
 @Getter
-@Data
+@Builder
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @AllArgsConstructor(access = AccessLevel.PROTECTED)
 public class SignupFormDTO {
