+++ src/main/java/com/ssafy/solvedpick/accounts/dto/TokenResponse.java
@@ -3,7 +3,6 @@ package com.ssafy.solvedpick.accounts.dto;
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
-import lombok.Data;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
