+++ src/main/java/com/ssafy/solvedpick/members/repository/MemberRepository.java
@@ -1,10 +1,10 @@
-package com.ssafy.solvedpick.accounts.repository;
-
-import com.ssafy.solvedpick.accounts.domain.Member;
+package com.ssafy.solvedpick.members.repository;
 
 import java.util.Optional;
 
 import org.springframework.data.jpa.repository.JpaRepository;
+
+import com.ssafy.solvedpick.members.domain.Member;
 
 public interface MemberRepository extends JpaRepository<Member, Long>{
 	boolean existsByUsername(String username);
