+++ src/main/java/com/ssafy/solvedpick/ownedavatar/service/impl/OwnedAvatarServiceImpl.java
@@ -1,7 +1,7 @@
 package com.ssafy.solvedpick.ownedavatar.service.impl;
 
-import com.ssafy.solvedpick.accounts.domain.Member;
 import com.ssafy.solvedpick.avatars.domain.Grade;
+import com.ssafy.solvedpick.members.domain.Member;
 import com.ssafy.solvedpick.ownedavatar.domain.OwnedAvatar;
 import com.ssafy.solvedpick.ownedavatar.dto.OwnedAvatarDTO;
 import com.ssafy.solvedpick.ownedavatar.repository.OwnedAvatarRepository;
